# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_diogomcd_fake-mill-api_internal_models.Bank:
    properties:
      code:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - code
    - name
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.BankAccountResponse:
    properties:
      account:
        maxLength: 15
        minLength: 5
        type: string
      accountType:
        enum:
        - checking
        - savings
        type: string
      agency:
        maxLength: 6
        minLength: 4
        type: string
      bank:
        $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.Bank'
    required:
    - account
    - accountType
    - agency
    - bank
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.CNPJResponse:
    properties:
      cnpj:
        type: string
      valid:
        type: boolean
    required:
    - cnpj
    - valid
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.CNPJValidationResponse:
    properties:
      cnpj:
        type: string
      valid:
        type: boolean
    required:
    - cnpj
    - valid
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.CPFResponse:
    properties:
      cpf:
        type: string
      valid:
        type: boolean
    required:
    - cpf
    - valid
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.CPFValidationResponse:
    properties:
      cpf:
        type: string
      valid:
        type: boolean
    required:
    - cpf
    - valid
    type: object
...
  github_com_diogomcd_fake-mill-api_internal_models.CreditCardResponse:
    properties:
      brand:
        enum:
        - Visa
        - Mastercard
        - Elo
        - Amex
        type: string
      cvv:
        type: string
      expirationDate:
        type: string
      holderName:
        maxLength: 100
        minLength: 5
        type: string
      number:
        type: string
    required:
    - brand
    - cvv
    - expirationDate
    - holderName
    - number
    type: object
...
  github_com_diogomcd_fake-mill-api_internal_models.PhoneValidationResponse:
    properties:
      country_code:
        type: string
      ddi:
        type: integer
      e164_format:
        type: string
      international_format:
        type: string
      national_format:
        type: string
      number_type:
        enum:
        - MOBILE
        - FIXED_LINE
        - FIXED_LINE_OR_MOBILE
        type: string
      phone_number:
        type: string
      valid:
        type: boolean
    required:
    - country_code
    - ddi
    - e164_format
    - international_format
    - national_format
    - number_type
    - phone_number
    - valid
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.RGResponse:
    properties:
      expirationDate:
        type: string
      issueDate:
        type: string
      issuer:
        maxLength: 10
        minLength: 3
        type: string
      rg:
        type: string
      state:
        type: string
    required:
    - expirationDate
    - issueDate
    - issuer
    - rg
    - state
    type: object
  github_com_diogomcd_fake-mill-api_internal_models.RGValidationResponse:
    properties:
      rg:
        type: string
      valid:
        type: boolean
    required:
    - rg
    - valid
    type: object
...
host: fakemill.com
info:
  contact: {}
  description: API REST para geração de dados fake brasileiros para testes, com foco
    em performance e simplicidade.
  license:
    name: GPL v3
    url: https://www.gnu.org/licenses/gpl-3.0.html
  termsOfService: http://swagger.io/terms/
  title: Fake Mill API
  version: 1.0.0
paths:
...
  /bank-account:
    get:
      consumes:
      - application/json
      description: Gera dados de uma conta bancária fictícia, incluindo banco, agência
        e número da conta.
      parameters:
      - default: 1
        description: Quantidade de contas (1-200)
        in: query
        maximum: 200
        minimum: 1
        name: quantity
        type: integer
      - description: 'Código do banco (ex: 001, 237)'
        in: query
        name: bank
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.BankAccountResponse'
            type: array
      summary: Gera dados bancários fictícios
      tags:
      - Financeiro
  /cnpj:
    get:
      consumes:
      - application/json
      description: Gera um ou mais números de CNPJ válidos ou inválidos, com opção
        de formatação.
      parameters:
      - default: 1
        description: Quantidade de CNPJs (1-200)
        in: query
        maximum: 200
        minimum: 1
        name: quantity
        type: integer
      - default: true
        description: Retorna formatado (XX.XXX.XXX/XXXX-XX)
        in: query
        name: formatted
        type: boolean
      - default: true
        description: Gera CNPJ válido com dígitos verificadores corretos
        in: query
        name: valid
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.CNPJResponse'
            type: array
      summary: Gera CNPJ válido ou inválido
      tags:
      - Documentos
...
  /cpf:
    get:
      consumes:
      - application/json
      description: Gera um ou mais números de CPF válidos ou inválidos, com opção
        de formatação.
      parameters:
      - default: 1
        description: Quantidade de CPFs (1-200)
        in: query
        maximum: 200
        minimum: 1
        name: quantity
        type: integer
      - default: true
        description: Retorna formatado (XXX.XXX.XXX-XX)
        in: query
        name: formatted
        type: boolean
      - default: true
        description: Gera CPF válido com dígitos verificadores corretos
        in: query
        name: valid
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.CPFResponse'
            type: array
      summary: Gera CPF válido ou inválido
      tags:
      - Documentos
  /credit-card:
    get:
      consumes:
      - application/json
      description: Gera dados de um cartão de crédito fictício, incluindo número,
        bandeira, CVV e data de validade.
      parameters:
      - default: 1
        description: Quantidade de cartões (1-200)
        in: query
        maximum: 200
        minimum: 1
        name: quantity
        type: integer
      - description: Bandeira do cartão
        enum:
        - visa
        - mastercard
        - elo
        - amex
        in: query
        name: brand
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.CreditCardResponse'
            type: array
      summary: Gera dados de cartão de crédito fictício
      tags:
      - Financeiro
...
  /rg:
    get:
      consumes:
      - application/json
      description: Gera um ou mais números de RG válidos ou inválidos.
      parameters:
      - default: 1
        description: Quantidade de RGs (1-200)
        in: query
        maximum: 200
        minimum: 1
        name: quantity
        type: integer
      - default: true
        description: Retorna formatado (XX.XXX.XXX-X)
        in: query
        name: formatted
        type: boolean
      - description: 'UF do estado (ex: SP, RJ)'
        in: query
        name: state
        type: string
      - default: true
        description: Gera RG válido com dígito verificador correto
        in: query
        name: valid
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.RGResponse'
            type: array
      summary: Gera RG válido ou inválido
      tags:
      - Documentos
  /validate/cnpj/{cnpj}:
    get:
      consumes:
      - application/json
      description: Verifica se um número de CNPJ é válido de acordo com o algoritmo
        oficial.
      parameters:
      - description: Número de CNPJ a validar (com ou sem formatação)
        in: path
        name: cnpj
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.CNPJValidationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Valida CNPJ
      tags:
      - Validação
  /validate/cpf/{cpf}:
    get:
      consumes:
      - application/json
      description: Verifica se um número de CPF é válido de acordo com o algoritmo
        oficial.
      parameters:
      - description: Número de CPF a validar (com ou sem formatação)
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.CPFValidationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Valida CPF
      tags:
      - Validação
  /validate/phone:
    get:
      consumes:
      - application/json
      description: Valida um número de telefone e retorna informações detalhadas sobre
        ele.
      parameters:
      - description: Número de telefone a validar (com ou sem DDI)
        in: query
        name: phone_number
        required: true
        type: string
      - description: 'Código do país ISO 3166-1 alpha-2 (ex: BR)'
        in: query
        name: country_code
        type: string
      - description: 'DDI - código de discagem internacional (ex: 55)'
        in: query
        name: ddi
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.PhoneValidationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Valida Telefone
      tags:
      - Validação
  /validate/rg/{rg}:
    get:
      consumes:
      - application/json
      description: Verifica se um número de RG tem o formato básico correto.
      parameters:
      - description: Número de RG a validar (com ou sem formatação)
        in: path
        name: rg
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_diogomcd_fake-mill-api_internal_models.RGValidationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Valida RG
      tags:
      - Validação
...
schemes:
- http
- https
swagger: "2.0"
...